fixed lazy tests
fix namespace body style
renamed ProxyScheduledSubject
add ProxyScheduledSubject tests
