+++ src/ReactiveMarbles.Mvvm.Tests/ReactiveObjectFixture.cs
@@ -8,113 +8,112 @@ using System.Runtime.Serialization;
 using DynamicData.Binding;
 using ReactiveUI;
 
-namespace ReactiveMarbles.Mvvm.Tests
+namespace ReactiveMarbles.Mvvm.Tests;
+
+/// <summary>
+/// A test fixture.
+/// </summary>
+/// <seealso cref="RxObject" />
+[DataContract]
+public class ReactiveObjectFixture : ReactiveObject
 {
+    [IgnoreDataMember]
+    private string? _isNotNullString;
+
+    [IgnoreDataMember]
+    private string? _isOnlyOneWord;
+
+    private string? _notSerialized;
+
+    [IgnoreDataMember]
+    private int? _nullableInt;
+
+    [IgnoreDataMember]
+    private string? _pocoProperty;
+
+    [IgnoreDataMember]
+    private List<string>? _stackOverflowTrigger;
+
+    [IgnoreDataMember]
+    private string? _usesExprRaiseSet;
+
     /// <summary>
-    /// A test fixture.
+    /// Initializes a new instance of the <see cref="ReactiveObjectFixture"/> class.
     /// </summary>
-    /// <seealso cref="RxObject" />
-    [DataContract]
-    public class ReactiveObjectFixture : ReactiveObject
+    public ReactiveObjectFixture() => TestCollection = new ObservableCollectionExtended<int>();
+
+    /// <summary>
+    /// Gets or sets the is not null string.
+    /// </summary>
+    [DataMember]
+    public string? IsNotNullString
     {
-        [IgnoreDataMember]
-        private string? _isNotNullString;
+        get => _isNotNullString;
+        set => this.RaiseAndSetIfChanged(ref _isNotNullString, value);
+    }
 
-        [IgnoreDataMember]
-        private string? _isOnlyOneWord;
+    /// <summary>
+    /// Gets or sets the is only one word.
+    /// </summary>
+    [DataMember]
+    public string? IsOnlyOneWord
+    {
+        get => _isOnlyOneWord;
+        set => this.RaiseAndSetIfChanged(ref _isOnlyOneWord, value);
+    }
 
-        private string? _notSerialized;
+    /// <summary>
+    /// Gets or sets the not serialized.
+    /// </summary>
+    public string? NotSerialized
+    {
+        get => _notSerialized;
+        set => this.RaiseAndSetIfChanged(ref _notSerialized, value);
+    }
 
-        [IgnoreDataMember]
-        private int? _nullableInt;
+    /// <summary>
+    /// Gets or sets the nullable int.
+    /// </summary>
+    [DataMember]
+    public int? NullableInt
+    {
+        get => _nullableInt;
+        set => this.RaiseAndSetIfChanged(ref _nullableInt, value);
+    }
 
-        [IgnoreDataMember]
-        private string? _pocoProperty;
+    /// <summary>
+    /// Gets or sets the poco property.
+    /// </summary>
+    [DataMember]
+    public string? PocoProperty
+    {
+        get => _pocoProperty;
+        set => _pocoProperty = value;
+    }
 
-        [IgnoreDataMember]
-        private List<string>? _stackOverflowTrigger;
+    /// <summary>
+    /// Gets or sets the stack overflow trigger.
+    /// </summary>
+    [DataMember]
+    public List<string>? StackOverflowTrigger
+    {
+        get => _stackOverflowTrigger;
+        set => this.RaiseAndSetIfChanged(ref _stackOverflowTrigger, value?.ToList());
+    }
 
-        [IgnoreDataMember]
-        private string? _usesExprRaiseSet;
+    /// <summary>
+    /// Gets or sets the test collection.
+    /// </summary>
+    [DataMember]
+    public ObservableCollectionExtended<int> TestCollection { get; protected set; }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ReactiveObjectFixture"/> class.
-        /// </summary>
-        public ReactiveObjectFixture() => TestCollection = new ObservableCollectionExtended<int>();
-
-        /// <summary>
-        /// Gets or sets the is not null string.
-        /// </summary>
-        [DataMember]
-        public string? IsNotNullString
-        {
-            get => _isNotNullString;
-            set => this.RaiseAndSetIfChanged(ref _isNotNullString, value);
-        }
-
-        /// <summary>
-        /// Gets or sets the is only one word.
-        /// </summary>
-        [DataMember]
-        public string? IsOnlyOneWord
-        {
-            get => _isOnlyOneWord;
-            set => this.RaiseAndSetIfChanged(ref _isOnlyOneWord, value);
-        }
-
-        /// <summary>
-        /// Gets or sets the not serialized.
-        /// </summary>
-        public string? NotSerialized
-        {
-            get => _notSerialized;
-            set => this.RaiseAndSetIfChanged(ref _notSerialized, value);
-        }
-
-        /// <summary>
-        /// Gets or sets the nullable int.
-        /// </summary>
-        [DataMember]
-        public int? NullableInt
-        {
-            get => _nullableInt;
-            set => this.RaiseAndSetIfChanged(ref _nullableInt, value);
-        }
-
-        /// <summary>
-        /// Gets or sets the poco property.
-        /// </summary>
-        [DataMember]
-        public string? PocoProperty
-        {
-            get => _pocoProperty;
-            set => _pocoProperty = value;
-        }
-
-        /// <summary>
-        /// Gets or sets the stack overflow trigger.
-        /// </summary>
-        [DataMember]
-        public List<string>? StackOverflowTrigger
-        {
-            get => _stackOverflowTrigger;
-            set => this.RaiseAndSetIfChanged(ref _stackOverflowTrigger, value?.ToList());
-        }
-
-        /// <summary>
-        /// Gets or sets the test collection.
-        /// </summary>
-        [DataMember]
-        public ObservableCollectionExtended<int> TestCollection { get; protected set; }
-
-        /// <summary>
-        /// Gets or sets the uses expr raise set.
-        /// </summary>
-        [DataMember]
-        public string? UsesExprRaiseSet
-        {
-            get => _usesExprRaiseSet;
-            set => this.RaiseAndSetIfChanged(ref _usesExprRaiseSet, value);
-        }
+    /// <summary>
+    /// Gets or sets the uses expr raise set.
+    /// </summary>
+    [DataMember]
+    public string? UsesExprRaiseSet
+    {
+        get => _usesExprRaiseSet;
+        set => this.RaiseAndSetIfChanged(ref _usesExprRaiseSet, value);
     }
 }
