+++ src/ReactiveMarbles.Mvvm/ILoggable.cs
@@ -2,13 +2,12 @@
 // ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-namespace ReactiveMarbles.Mvvm
+namespace ReactiveMarbles.Mvvm;
+
+/// <summary>
+/// Marker interface for classes that want access to the underlying framework logger.
+/// </summary>
+public interface ILoggable
 {
-    /// <summary>
-    /// Marker interface for classes that want access to the underlying framework logger.
-    /// </summary>
-    public interface ILoggable
-    {
-        // TODO: [rlittlesii: October 11, 2021] US MS Logging to log MVVM internals.
-    }
+    // TODO: [rlittlesii: October 11, 2021] US MS Logging to log MVVM internals.
 }
