+++ src/ReactiveMarbles.Mvvm/ReactiveMarbles.Mvvm.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-      <TargetFrameworks>netstandard2.0;net5.0;net6.0</TargetFrameworks>
+      <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
       <LangVersion>latest</LangVersion>
       <PackageDescription>A MVVM framework that integrates with the Reactive Extensions for .NET to create elegant, testable MVVM components that run on any mobile or desktop platform. This is the base package with the base platform implementations</PackageDescription>
       <Nullable>enable</Nullable>
@@ -9,6 +9,7 @@
 
     <ItemGroup>
       <PackageReference Include="DynamicData" />
+      <PackageReference Include="ReactiveMarbles.Extensions" />
       <PackageReference Include="ReactiveMarbles.Locator" />
       <PackageReference Include="System.Reactive" />
     </ItemGroup>
