+++ src/ReactiveMarbles.Mvvm/RxDisposableObject.cs
@@ -4,26 +4,25 @@
 
 using System;
 
-namespace ReactiveMarbles.Mvvm
-{
-    /// <summary>
-    /// Rx object that extends the <see cref="IDisposable"/> interface.
-    /// </summary>
-    public class RxDisposableObject : RxObject, IDisposable
-    {
-        /// <inheritdoc />
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
+namespace ReactiveMarbles.Mvvm;
 
-        /// <summary>
-        /// Disposes of the resources.
-        /// </summary>
-        /// <param name="disposing">Disposing.</param>
-        protected virtual void Dispose(bool disposing)
-        {
-        }
+/// <summary>
+/// Rx object that extends the <see cref="IDisposable"/> interface.
+/// </summary>
+public class RxDisposableObject : RxObject, IDisposable
+{
+    /// <inheritdoc />
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    /// <summary>
+    /// Disposes of the resources.
+    /// </summary>
+    /// <param name="disposing">Disposing.</param>
+    protected virtual void Dispose(bool disposing)
+    {
     }
 }
