+++ src/directory.packages.props
@@ -21,6 +21,7 @@
 
   <ItemGroup>
     <PackageVersion Include="DynamicData" Version="7.4.9" />
+    <PackageVersion Include="ReactiveMarbles.Extensions" Version="1.0.1" />
     <PackageVersion Include="ReactiveMarbles.Locator" Version="1.1.4" />
     <PackageVersion Include="ReactiveMarbles.PropertyChanged" Version="2.0.27" />
     <PackageVersion Include="System.Reactive" Version="5.0.0" />
